+++ rog-aura/src/lib.rs
@@ -126,7 +126,5 @@ pub enum PowerZones {
     Lid = 3,
     /// The led strip on the rear of some laptops
     RearGlow = 4,
-    /// On pre-2021 laptops there is either 1 or 2 zones used
-    KeyboardAndLightbar = 5,
     None = 255,
 }